chore: 서비스 변경, 엔티티 업데이트

+++ src/main/java/com/book_everywhere/service/ReviewService.java
@@ -4,11 +4,11 @@ import com.book_everywhere.domain.review.Review;
 import com.book_everywhere.domain.review.ReviewRepository;
 import com.book_everywhere.web.dto.review.ReviewDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,37 +21,44 @@ public class ReviewService {
 
     //등록
     @Transactional
-    public Review createReview(ReviewDto reviewDto) {
-        Review review = new Review().createReview(reviewDto);
-        review.setIsPrivate(reviewDto.getIsPrivate());
-        return reviewRepository.save(review);
+    public void createReview(ReviewDto reviewDto) {
+        Review review = new Review().createFromDto(reviewDto);
+        reviewRepository.save(review);
     }
 
 
     //수정
     @Transactional
-    public Review updateReview(Long id, ReviewDto reviewDto) {
+    public void updateReview(Long id, ReviewDto reviewDto) {
         Review review = findOneReview(id);
-        Review updateReview = review.update(reviewDto);
-        return reviewRepository.save(updateReview);
+        review.update(reviewDto);
     }
-
 
 
     //삭제
     @Transactional
     public void deleteReview(Long id) {
+        if (!reviewRepository.existsById(id)) {
+            throw new IllegalArgumentException("Review does not exist");
+        }
         reviewRepository.deleteById(id);
     }
 
 
     //조회
-    public List<Review> findReviews() {
-        return reviewRepository.findAll();
+    //공유 지도에서의 조회
+    public List<Review> findPublicReviews(boolean isPrivate, Pageable pageable) {
+        return reviewRepository.findByIsPrivate(isPrivate, pageable);
     }
 
+    //아이디 값 하나만 조회
     public Review findOneReview(Long id) {
         return reviewRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("Review does not exist"));
+    }
+
+    //테스트용
+    public List<Review> findReviews() {
+        return reviewRepository.findAll();
     }
 }
